% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggcov.R
\name{get_detailed_pathway_info}
\alias{get_detailed_pathway_info}
\title{Get Detailed Pathway Information for KOs}
\usage{
get_detailed_pathway_info(ko_list)
}
\arguments{
\item{ko_list}{A character vector of KO identifiers.}
}
\value{
A data frame with columns: KO, Pathway_ID, Pathway_Description.
}
\description{
Retrieves detailed metabolic pathway information for a list of KOs.
}
\examples{

library(keggcov)
my_kos <- c("K00844", "K00164")
df <- get_detailed_pathway_info(my_kos)
head(df)

}
