% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggcov.R
\name{get_pathway_covariance}
\alias{get_pathway_covariance}
\title{Create a Pathway-Based Covariance Matrix from KEGG orthologs (KOs)}
\usage{
get_pathway_covariance(ko_list, strict_metabolic = FALSE)
}
\arguments{
\item{ko_list}{A character vector of KO identifiers (e.g., c("K00844", "K01803"))}

\item{strict_metabolic}{Logical. If TRUE, limits analysis to only manually drawn reference metabolic pathways ('map' prefix).}
}
\value{
A list containing the covariance matrix and the binary pathway profile matrix.
}
\description{
Create a pathway-based variance-covariance matrix from a set of KEGG orthologs (KOs).
}
\details{
This function leverages the KEGG metabolic pathway database \url{https://www.genome.jp/kegg/pathway.html} to estimate a variance-covariance structure for sets of KOs identified by high-throughput sequencing. The output matrix can be interpreted through its (1) Diagonal Elements (Variance), which represent the functional versatility of the KO (i.e., a high variance on the diagonal means the KO is involved in many different pathways (a "hub" gene); a low variance means the KO is specialized to only one or two specific pathways); and through its (2) Off-Diagonal Elements (Covariance), which prepresent pathway partnerships. Positive high covariance indicates that the KOs are "pathway partners." They almost always appear together (e.g., subunits of the same complex or sequential enzymes in a linear pathway). Zeros indicated that the KOs share no biological processes (e.g., a DNA replication gene vs. a glycolysis gene).
}
\examples{

# Define KOs
# K00844, K00845: Hexokinase/Glucokinase (Glycolysis)
# K00164: Pyruvate Dehydrogenase (TCA Cycle)
# K00500: Fatty Acid Synthase (Fatty Acid Biosynthesis)
my_kos <- c("K00844", "K00845", "K00164", "K00500")

# Run the function, without metabolic filtering

library(keggcov)

results <- get_pathway_covariance(my_kos)

str(results)
rownames(results$binary_matrix)

if (!is.null(results)) {
    # Print the Covariance Matrix (Rounded for readability)
    print(round(results$cov_matrix, 3))

    # Visualization
    heatmap(results$cov_matrix,
        main = "Pathway Co-Membership Covariance",
        symm = TRUE,
        col = cm.colors(256)
    )
}

# Run the function, with metabolic filtering
results <- get_pathway_covariance(my_kos, strict_metabolic = TRUE)

str(results)

if (!is.null(results)) {
    # Print the Covariance Matrix (Rounded for readability)
    print(round(results$cov_matrix, 3))

    # Visualization
    heatmap(results$cov_matrix,
        main = "Metabolic Pathway Co-Membership Covariance",
        symm = TRUE,
        col = cm.colors(256)
    )
}

}
